+++ projects/ids-enterprise-ng/src/lib/pager/soho-standalone-pager.component.ts
@@ -99,6 +99,11 @@ export class SohoStandalonePagerComponent implements AfterViewInit, AfterViewChe
     this.updateRequired = !!this.pager;
   }
 
+  @Input() set pageSizeMenuSettings(options: object) {
+    this.options.pageSizeMenuSettings = options;
+    this.updateRequired = !!this.pager;
+  }
+
   @Output() firstPage: EventEmitter<Object> = new EventEmitter<Object>();
   @Output() lastPage: EventEmitter<Object> = new EventEmitter<Object>();
   @Output() previousPage: EventEmitter<Object> = new EventEmitter<Object>();