+++ src/app/autocomplete/autocomplete.demo.ts
@@ -13,17 +13,58 @@ import { SohoAutoCompleteComponent } from 'ids-enterprise-ng';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AutocompleteDemoComponent implements AfterViewInit {
-  public url = 'http://localhost:4200/app/demodata/cities.demo.json?term=';
-  public statesUrl = 'http://localhost:4200/app/demodata/states.demo.json?term=';
+  public url = './app/demodata/cities.demo.json?term=';
+  public statesUrl = './app/demodata/states.demo.json?term=';
   public states = [
+    'Alabama',
     'Alaska',
     'Arizona',
+    'Arkansas',
     'California',
     'Colorado',
+    'Connecticut',
+    'Delaware',
+    'Florida',
+    'Georgia',
+    'Hawaii',
+    'Idaho',
+    'Illinois',
+    'Indiana',
+    'Iowa',
+    'Kansas',
+    'Kentucky',
+    'Louisiana',
+    'Maine',
+    'Maryland',
+    'Massachusetts',
+    'Michigan',
     'Minnesota',
+    'Mississippi',
+    'Missouri',
+    'Montana',
+    'Nebraska',
+    'Nevada',
+    'New Hampshire',
+    'New Jersey',
+    'New Mexico',
+    'New York',
+    'North Carolina',
     'North Dakota',
+    'Ohio',
+    'Oklahoma',
     'Oregon',
+    'Pennsylvania',
+    'Rhode Island',
+    'South Carolina',
+    'South Dakota',
+    'Tennessee',
+    'Texas',
+    'Utah',
+    'Vermont',
+    'Virginia',
     'Washington',
+    'West Virginia',
+    'Wisconsin',
     'Wyoming'
   ];
   public xssSource: SohoAutoCompleteSource = [